style the movie list and make it clickable

+++ utils/dateFormater.js
@@ -1,5 +1,5 @@
 export const dateFormatter = (dateString) => {
 	const date = new Date(dateString);
-	const options = { year: "numeric", month: "long", day: "numeric" };
+	const options = { year: "numeric", month: "short", day: "numeric" };
 	return date.toLocaleDateString("en-US", options);
 };
